requirements.txt
flask
pdfplumber
python-docx
Werkzeug
google-generativeai
fpdf
python-dotenv
cloudinary
